Move repeated methods to Conversion
